setSetpoint to Encoder when nothing is running

+++ src/main/java/frc/robot/commands/PositionsCommand.java
@@ -7,10 +7,10 @@ import frc.robot.subsystems.ElevatorSubsystem;
 public class PositionsCommand extends CommandBase{
     ElevatorSubsystem elevator;
     double setSetPoint;
-    int setpoint;
+    double setpoint;
     String elevatorPosition;
 
-     public PositionsCommand(ElevatorSubsystem elevatorSubsystem, int setpoint){
+     public PositionsCommand(ElevatorSubsystem elevatorSubsystem, double setpoint){
         elevator = elevatorSubsystem;
         this.setpoint = setpoint;
         addRequirements(elevator);
